Improve compatibility with twemproxy
This commit makes three changes to improve the compatibility with
twemproxy. It removes the concept of the "unknown" connection state, and
instead assumes persistent connections are active, preferring to retry
automatically. It adds a value in php.ini for specifying the number of
retry attempts and the timeout value. Finally, it removes the "version"
command completely and changes the delete command to comply with the
latest memcache protocol standard.